+++ src/PSDocs/Models/Document.cs
@@ -1,4 +1,5 @@
-﻿using System.Collections.Specialized;
+﻿
+using System.Collections.Specialized;
 
 namespace PSDocs.Models
 {