+++ src/PSDocs/Resources/PSDocsResources.Designer.cs
@@ -79,6 +79,15 @@ namespace PSDocs.Resources {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to A culture must be specified when no culture is specified by the system. To specify a culture use the &apos;Output.Culture&apos; option..
+        /// </summary>
+        internal static string InvariantCulture {
+            get {
+                return ResourceManager.GetString("InvariantCulture", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Options file does not exist..
         /// </summary>
         internal static string OptionsNotFound {