+++ CHANGELOG.md
@@ -6,6 +6,8 @@ What's changed since pre-release v0.7.0-B2008035:
 
 - Engine features:
   - Added support for localized strings using the `$LocalizedData` variable. [#91](https://github.com/BernieWhite/PSDocs/issues/91)
+- Bug fixes:
+  - Fixed inconsistencies with default options file name. [#103](https://github.com/BernieWhite/PSDocs/issues/103)
 
 ## v0.7.0-B2008035 (pre-release)